requirements.txt
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
chromedriver-binary==74.0.3729.6.0
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
exceptiongroup==1.3.0
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.1
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pillow==11.3.0
preshed==3.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
seaborn==0.13.2
selenium==4.34.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
textblob==0.19.0
thinc==8.3.6
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
tweety==0.1.6
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vaderSentiment==3.3.2
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
tweepy==4.16.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blis==1.3.0
blurhash==1.1.4
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
chromedriver-binary==74.0.3729.6.0
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
decorator==5.2.1
exceptiongroup==1.3.0
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
Mastodon.py==2.0.1
matplotlib==3.10.3
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pillow==11.3.0
preshed==3.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
pytz==2025.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
seaborn==0.13.2
selenium==4.34.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
textblob==0.19.0
thinc==8.3.6
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
tweepy==4.16.0
tweety==0.1.6
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vaderSentiment==3.3.2
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blis==1.3.0
blurhash==1.1.4
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
chromedriver-binary==74.0.3729.6.0
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
decorator==5.2.1
exceptiongroup==1.3.0
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
Mastodon.py==2.0.1
matplotlib==3.10.3
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pillow==11.3.0
praw==7.8.1
prawcore==2.4.0
preshed==3.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
pytz==2025.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
seaborn==0.13.2
selenium==4.34.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
textblob==0.19.0
thinc==8.3.6
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
tweepy==4.16.0
tweety==0.1.6
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
vaderSentiment==3.3.2
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
